EKIR-628 Rename SelectBook to SelectBookCard
